chore(conv rendering): rename file and flatten dir structure
