+++ front/lib/api/assistant/conversation_rendering/helpers.ts
